requirements.txt
docx==0.2.4
Pillow==9.2.0
pytesseract==0.3.10
